Move serde functions out of their own module
