+++ Makefile
@@ -1,5 +1,5 @@
 CFLAGS = -Wall -g -O2
-SRCS = hazmat.c randombytes.c serialize.c sss.c keccak.c tweetnacl.c
+SRCS = hazmat.c randombytes.c sss.c keccak.c tweetnacl.c
 OBJS := ${SRCS:.c=.o}
 
 all: libsss.a
