+++ hazmat.h
@@ -15,6 +15,9 @@
 #include <inttypes.h>
 
 
+#define sss_KEYSHARE_SERIALIZED_LEN 33 /* 1 + 32 */
+
+
 /*
  * One share of a cryptographic key which is shared using Shamir's
  * the `sss_create_keyshares` function.
@@ -23,6 +26,22 @@ typedef struct {
 	uint8_t x;
 	uint8_t y[32];
 } sss_Keyshare;
+
+
+/*
+ * Serialize a `sss_Keyshare` struct into a bytestring pointed to by `out`.
+ * The resulting bytestring will be `sss_KEYSHARE_SERIALIZED_LEN` long. The
+ * caller must make sure that `out` is large enough to contain this data.
+ */
+void sss_serialize_keyshare(uint8_t *out, const sss_Keyshare *keyshare);
+
+
+/*
+ * Parse the bytestring `in` that has previously been generated by the
+ * `sss_serialize_keyshare` function. This function will write the resulting
+ * `sss_Keyshare` struct to the location pointed to by `keyshare`.
+ */
+void sss_deserialize_keyshare(sss_Keyshare *keyshare, const uint8_t *in);
 
 
 /*
