+++ sss.c
@@ -36,6 +36,30 @@ static const unsigned char nonce[crypto_secretbox_NONCEBYTES] = { 0 };
 
 
 /*
+ * Serialize a `sss_Share` struct into a bytestring pointed to by `out`. The
+ * resulting bytestring will be `sss_SHARE_SERIALIZED_LEN` long. The caller must
+ * make sure that `out` is large enough to contain this data.
+ */
+static void sss_serialize_share(uint8_t *out, const sss_Share *share)
+{
+	sss_serialize_keyshare(out, &share->keyshare);
+	memcpy(&out[sss_KEYSHARE_SERIALIZED_LEN], &share->c, sss_CLEN);
+}
+
+
+/*
+ * Parse the bytestring `in` that has previously been generated by the
+ * `sss_serialize_share` function. This function will write the resulting
+ * `sss_Share` struct to the location pointed to by `share`.
+ */
+static void sss_deserialize_share(sss_Share *share, const uint8_t *in)
+{
+	sss_unserialize_keyshare(&share->keyshare, in);
+	memcpy(&share->c, &in[sss_KEYSHARE_SERIALIZED_LEN], sss_CLEN);
+}
+
+
+/*
  * Create `n` shares with theshold `k` and write them to `out`
  */
 void sss_create_shares(sss_Share *out, const unsigned char *data,
