+++ sss.h
@@ -26,6 +26,9 @@ Length of the message (must be known at compile-time)
 #define sss_CLEN (sss_MLEN + 16)
 
 
+#define sss_SHARE_SERIALIZED_LEN (sss_CLEN + sss_KEYSHARE_SERIALIZED_LEN)
+
+
 /*
  * One share of a secret which is shared using Shamir's
  * the `sss_create_shares` function.
